Optimize Status.as_home_timeline
This solves a performance regression introduced by commit
bab5a18232a163b0c3c6a245f7f95d50d7022b36.
