+++ app/controllers/api/v1/timelines/tag_controller.rb
@@ -31,16 +31,15 @@ class Api::V1::Timelines::TagController < Api::BaseController
     if @tag.nil?
       []
     else
-      tag_timeline_statuses.paginate_by_max_id(
-        limit_param(DEFAULT_STATUSES_LIMIT),
-        params[:max_id],
-        params[:since_id]
+      Status.as_tag_timeline(
+        @tag,
+        account: current_account,
+        local_only: params[:local],
+        limit: limit_param(DEFAULT_STATUSES_LIMIT),
+        max_id: params[:max_id],
+        since_id: params[:since_id]
       )
     end
-  end
-
-  def tag_timeline_statuses
-    Status.as_tag_timeline(@tag, current_account, params[:local])
   end
 
   def insert_pagination_headers
