+++ app/controllers/tags_controller.rb
@@ -5,7 +5,15 @@ class TagsController < ApplicationController
 
   def show
     @tag      = Tag.find_by!(name: params[:id].downcase)
-    @statuses = @tag.nil? ? [] : Status.as_tag_timeline(@tag, current_account, params[:local]).paginate_by_max_id(20, params[:max_id])
+    @statuses = if @tag.nil?
+                  []
+                else
+                  Status.as_tag_timeline(@tag,
+                                         account: current_account,
+                                         local_only: params[:local],
+                                         limit: 20,
+                                         max_id: params[:max_id])
+                end
     @statuses = cache_collection(@statuses, Status)
   end
 end
