+++ app/models/feed.rb
@@ -25,8 +25,7 @@ class Feed
   end
 
   def from_database(limit, max_id, since_id)
-    Status.as_home_timeline(@account)
-          .paginate_by_max_id(limit, max_id, since_id)
+    Status.as_home_timeline(@account, limit: limit, max_id: max_id, since_id: since_id)
           .reject { |status| FeedManager.instance.filter?(:home, status, @account.id) }
   end
